+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Constellation Changes
 
 ## Changes in July 2024
--   Moved hashmod package from Core Graph Utilities to Core Import Export Plugins
+-   Moved hashmod package from Core Graph Utilities to Core Import Export Plugins as a more appropriate module.
+-   Moved `NoGraphPane` from Layers View to the View Framework so that other views can use it.
+-   Updated `NoGraphPane` to take two parameters needed for the abstraction.
 
 ## Changes in May 2024
 -   Removed `FloatArray.clone()` and replaced with a constructor that takes a `FloatArray` object.